+++ ChangeLog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Version [0.5.0](https://github.com/cedx/html.net/compare/v0.4.0...v0.5.0)
+- Added the `New-ColGroupElement` cmdlets.
+- Added the `colgroup` aliases.
+
 ## Version [0.4.0](https://github.com/cedx/html.net/compare/v0.3.0...v0.4.0)
 - Added support for the `on*` and `tabindex` attributes.
 - Added the `New-AudioElement`, `New-BlockquoteElement`, `New-CanvasElement` and `New-ColElement` cmdlets.